# Speed up GetTerms